Header alignment default should be column alignment
If 'justify' is not included in header_props argument, the default
alignment should match the given column alignment. If the column is
left-justified, the header will be left-justified. If the column is
centered, the header will be centered. If no column alignment is
specified, the default (for both) is left-justified.